+++ karbon/ui/vmdlg_stroke.cc
@@ -89,7 +89,7 @@ VMDlgStroke::VMDlgStroke( KarbonPart *part ) : QTabDialog ( 0L, 0, true ), m_par
 	mOpacity = new KIntNumInput( 100, ogroupBox );
 	mOpacity->setRange( 0, 100, 1, true );
 	if( part->selection().count() > 0 ) // there is a selection, so take the opacity of first selected object
-		mOpacity->setValue( part->selection().getFirst()->stroke().color().opacity() * 100.0 );
+		mOpacity->setValue( static_cast<int>(part->selection().getFirst()->stroke().color().opacity() * 100.0) );
 	mainLayout->addWidget( ogroupBox, 2, 2 );
 
 	mainLayout->setSpacing( 2 );
